Update why.md
Mainly changing "Porquê" to "Por que", since "Porquê" is a noun, and should not be used to make questions. Also changed "ficheiro" to "arquivos" since it's a more common way to describe those. Changed some text that seemed to be auto-translated as they did not make much sense in PT and other because there were errors or inconsistencies.